+++ client/src/auth/RegisterModal.jsx
@@ -1,7 +1,7 @@
 import { useState, useEffect } from "react";
 import { Link, useNavigate } from "react-router";
 
-import { handleLoginOrRegister } from "/src/utils/userAccessUtils.js";
+import { handleLoginOrRegister } from "/src/utils/userUtils/userAccessUtils.js";
 
 import "/src/css/landingModal.css";
